+++ Plugin.BluetoothLE.Android.Tests/MainActivity.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Reflection;
+using Acr.UserDialogs;
+using Android;
 using Android.App;
+using Android.Content.PM;
 using Android.OS;
-using Plugin.BluetoothLE.Tests;
 using Xunit.Runners.UI;
-using Xunit.Sdk;
 
 
 namespace Plugin.BluetoothLE.Android.Tests
@@ -18,15 +19,34 @@ namespace Plugin.BluetoothLE.Android.Tests
     {
         protected override void OnCreate(Bundle bundle)
         {
-            this.AddExecutionAssembly(typeof(ExtensibilityPointFactory).Assembly);
-            this.AddTestAssembly(typeof(SpecificTests).Assembly);
+            GattConnectionConfig.DefaultConfiguration.AutoConnect = false;
+
+            this.RequestPermissions(new[]
+            {
+                Manifest.Permission.AccessCoarseLocation,
+                Manifest.Permission.BluetoothPrivileged
+            }, 0);
+
+            UserDialogs.Init(this);
+            //this.AddExecutionAssembly(typeof(ExtensibilityPointFactory).Assembly);
+            this.AddTestAssembly(typeof(BluetoothLE.Tests.DeviceTests).Assembly);
             this.AddTestAssembly(Assembly.GetExecutingAssembly());
 
-            this.AutoStart = true;
+            //CrossBleAdapter.AndroidUseNewScanner = false;
+            CrossBleAdapter.AndroidOperationPause = TimeSpan.FromMilliseconds(100);
+            //CrossBleAdapter.AndroidMaxAutoReconnectAttempts = 2;
+            //CrossBleAdapter.AndroidPerformActionsOnMainThread = false;
+
+            this.AutoStart = false;
             this.TerminateAfterExecution = false;
             //this.Writer =
 
             base.OnCreate(bundle);
+        }
+
+
+        public override void OnRequestPermissionsResult(int requestCode, string[] permissions, Permission[] grantResults)
+        {
         }
     }
 }