+++ Plugin.BluetoothLE.Shared/CrossBleAdapter.cs
@@ -29,17 +29,50 @@ namespace Plugin.BluetoothLE
 
 
         /// <summary>
-        /// Number of milliseconds to pause before service discovery (helps in combating GATT133 error) when service discovery is performed immediately after connection
+        /// Time span to pause before service discovery (helps in combating GATT133 error) when service discovery is performed immediately after connection
         /// DO NOT CHANGE this if you don't know what this is!
         /// </summary>
         public static TimeSpan AndroidPauseBeforeServiceDiscovery { get; set; } = TimeSpan.FromMilliseconds(750);
-        
+
+
+        /// <summary>
+        /// Disables the locking mechanism used by all read/writes within android plugin 
+        /// You must manage serial communication on your own without this - GOOD LUCK!
+        /// DO NOT CHANGE this if you don't know what this is!
+        /// </summary>
+        public static bool AndroidDisableLockMechanism { get; set; }
+
+
+        static TimeSpan? androidOpPause;
+
+        /// <summary>
+        /// Time span to pause android operations
+        /// DO NOT CHANGE this if you don't know what this is!
+        /// </summary>
+        public static TimeSpan? AndroidOperationPause
+        {
+            get
+            {
+                if (androidOpPause != null)
+                    return androidOpPause;
+
+                if (Android.OS.Build.VERSION.SdkInt < Android.OS.BuildVersionCodes.N)
+                    return TimeSpan.FromMilliseconds(150);
+
+                return null;
+            }
+            set => androidOpPause = value;
+        }
+
 
         /// <summary>
         /// Specifies the wait time before attempting an auto-reconnect
         /// DO NOT CHANGE if you don't know what this is!
         /// </summary>
         public static TimeSpan AndroidPauseBetweenAutoReconnectAttempts { get; set; } = TimeSpan.FromSeconds(1);
+
+
+        public static bool AndroidUseNewScanner { get; set; } = B.VERSION.SdkInt >= BuildVersionCodes.Lollipop;
 
 
         public static bool AndroidMainThreadSuggested =>