+++ Plugin.BluetoothLE.iOS.Tests/AppDelegate.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Reflection;
 using Foundation;
-using Plugin.BluetoothLE.Tests;
 using UIKit;
 using Xunit.Runner;
 using Xunit.Sdk;
@@ -15,10 +14,10 @@ namespace Plugin.BluetoothLE.iOS.Tests
         public override bool FinishedLaunching(UIApplication app, NSDictionary options)
         {
             this.AddExecutionAssembly(typeof(ExtensibilityPointFactory).Assembly);
-            this.AddTestAssembly(typeof(SpecificTests).Assembly);
+            this.AddTestAssembly(typeof(BluetoothLE.Tests.DeviceTests).Assembly);
             this.AddTestAssembly(Assembly.GetExecutingAssembly());
 
-            this.AutoStart = true;
+            this.AutoStart = false;
             this.TerminateAfterExecution = false;
 
             return base.FinishedLaunching(app, options);