+++ Samples/Samples.Uwp/MainPage.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Autofac;
 
 
 namespace Samples.Uwp
@@ -9,11 +8,7 @@ namespace Samples.Uwp
         public MainPage()
         {
             this.InitializeComponent();
-
-            var builder = new ContainerBuilder();
-            builder.RegisterModule(new PlatformModule());
-            var container = builder.Build();
-            this.LoadApplication(new Samples.App(container));
+            this.LoadApplication(new Samples.App());
         }
     }
 }